requirements.txt
djangorestframework-simplejwt
djangorestframework
django-cors-headers
